+++ src/BDSBunchSixTrack.cc
@@ -44,9 +44,9 @@ void BDSBunchSixTrack::SetOptions(const BDSParticleDefinition* beamParticle,
 				  const GMAD::Beam& beam,
 				  const BDSBunchType& distrType,
 				  G4Transform3D beamlineTransformIn,
-				  const G4double beamlineS)
+				  const G4double beamlineSIn)
 {
-  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineS);
+  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineSIn);
   fileName = G4String(beam.distrFile); 
   LoadSixTrackFile();
 }